+++ Adria/Graphics/GfxTexture.h
@@ -139,31 +139,31 @@ namespace adria
 		GfxTexture& operator=(GfxTexture&&) = delete;
 		~GfxTexture() = default;
 
-		[[maybe_unused]] size_t CreateSRV(GfxTextureSubresourceDesc const* desc = nullptr)
+		[[maybe_unused]] uint64 CreateSRV(GfxTextureSubresourceDesc const* desc = nullptr)
 		{
 			GfxTextureSubresourceDesc _desc = desc ? *desc : GfxTextureSubresourceDesc{};
 			return CreateSubresource(GfxSubresourceType_SRV, _desc);
 		}
-		[[maybe_unused]] size_t CreateUAV(GfxTextureSubresourceDesc const* desc = nullptr)
+		[[maybe_unused]] uint64 CreateUAV(GfxTextureSubresourceDesc const* desc = nullptr)
 		{
 			GfxTextureSubresourceDesc _desc = desc ? *desc : GfxTextureSubresourceDesc{};
 			return CreateSubresource(GfxSubresourceType_UAV, _desc);
 		}
-		[[maybe_unused]] size_t CreateRTV(GfxTextureSubresourceDesc const* desc = nullptr)
+		[[maybe_unused]] uint64 CreateRTV(GfxTextureSubresourceDesc const* desc = nullptr)
 		{
 			GfxTextureSubresourceDesc _desc = desc ? *desc : GfxTextureSubresourceDesc{};
 			return CreateSubresource(GfxSubresourceType_RTV, _desc);
 		}
-		[[maybe_unused]] size_t CreateDSV(GfxTextureSubresourceDesc const* desc = nullptr)
+		[[maybe_unused]] uint64 CreateDSV(GfxTextureSubresourceDesc const* desc = nullptr)
 		{
 			GfxTextureSubresourceDesc _desc = desc ? *desc : GfxTextureSubresourceDesc{};
 			return CreateSubresource(GfxSubresourceType_DSV, _desc);
 		}
 
-		ID3D11ShaderResourceView* SRV(size_t i = 0) const { return srvs[i].Get(); }
-		ID3D11UnorderedAccessView* UAV(size_t i = 0) const { return uavs[i].Get(); }
-		ID3D11RenderTargetView* RTV(size_t i = 0) const { return rtvs[i].Get(); }
-		ID3D11DepthStencilView* DSV(size_t i = 0) const { return dsvs[i].Get(); }
+		ID3D11ShaderResourceView* SRV(uint64 i = 0) const { return srvs[i].Get(); }
+		ID3D11UnorderedAccessView* UAV(uint64 i = 0) const { return uavs[i].Get(); }
+		ID3D11RenderTargetView* RTV(uint64 i = 0) const { return rtvs[i].Get(); }
+		ID3D11DepthStencilView* DSV(uint64 i = 0) const { return dsvs[i].Get(); }
 
 		ID3D11Resource* GetNative() const { return resource.Get(); }
 		ID3D11Resource* Detach() { return resource.Detach(); }
@@ -179,7 +179,7 @@ namespace adria
 		std::vector<ArcPtr<ID3D11DepthStencilView>> dsvs;
 
 	private:
-		[[maybe_unused]] size_t CreateSubresource(GfxSubresourceType type, GfxTextureSubresourceDesc const& view_desc)
+		[[maybe_unused]] uint64 CreateSubresource(GfxSubresourceType type, GfxTextureSubresourceDesc const& view_desc)
 		{
 			uint32 first_slice = view_desc.first_slice;
 			uint32 slice_count = view_desc.slice_count;
